Changes: 1) Implemented hierarchical numbering of Selectors (First-level selectors are numbered 1., 2., 3. etc.; Second level x.1, x.2, etc.; Third level x.y.1, x.y.2, etc...); 2) Separated input boxes into controlled components for greater flexibility.
TO-DO: Since the scraper at this point will only allow for the collection of text info, the type selection option is redundant. If a selector leads to other selectors, it is by definition a link. If it doesn't lead to other selectors, it represents a local information end-point. Therefore: a) the option box needs to be removed, while the type of selector needs to be adjusted dynamically with respect to its parent status (if it has children, it is a link).